+++ src/test/nextflow/co2footprint/CO2FootprintCLITest.groovy
@@ -17,7 +17,6 @@ class CO2FootprintCLITest extends  Specification {
 
     List<Path> outPaths = [
             outPath.resolve('trace_test.txt'),
-            outPath.resolve('summary_test.txt'),
             outPath.resolve('report_test.html'),
             outPath.resolve('provenance_test.json')
     ]